requirements.txt
# fastapi>=0.109.0
# python-dotenv>=1.0.0
# email-validator>=2.0.0
# SQLAlchemy>=2.0.0
# psycopg2-binary>=2.9.9
# requests>=2.32.0
# mangum>=0.14.0

# fastapi==0.95.2
# mangum==0.17.0
# pydantic==1.10.13
# sqlalchemy==1.4.54
# python-dotenv==1.0.1
# psycopg2-binary>=2.9.9

# fastapi==0.104.1
# mangum==0.17.0
# pydantic==2.5.0
# pydantic-core==2.14.5
# typing-extensions==4.8.0
# annotated-types==0.6.0
# exceptiongroup==1.2.0
# starlette==0.27.0
# anyio==3.7.1
# sniffio==1.3.0
# idna==3.6

fastapi==0.103.2
mangum>=0.17.0
pydantic==1.10.13
sqlalchemy==1.4.54
python-dotenv==1.0.1
psycopg2-binary>=2.9.9
email-validator>=2.0.0